Added action 'https' for upgrading to https
The action adds a rule for redirecting (301) to the https url, the
vhost registry now uses this to add a new listener on port 80 with
https upgrade enabled.
Also fixed names for ACL rules, they should be unique per rule
entry instead of fixed to the target group slug.
Path and host ACL are now separated.